+++ examples/react/src/components/Connected.tsx
@@ -26,7 +26,7 @@ import {
   useWriteContract
 } from 'wagmi'
 
-import { sponsoredContractAddresses } from '../config'
+import { sponsoredContractAddresses, getErc1155SaleContractConfig } from '../config'
 import { messageToSign } from '../constants'
 import { abi } from '../constants/nft-abi'
 import { delay, getCheckoutSettings, getOrderbookCalldata } from '../utils'
@@ -336,12 +336,14 @@ export const Connected = () => {
     // const salesContractAddress = '0xf0056139095224f4eec53c578ab4de1e227b9597'
     // const collectionAddress = '0x92473261f2c26f2264429c451f70b0192f858795'
     // const price = '200000000000000'
+    // const contractId = '674eb55a3d739107bbd18ecb'
 
     // // ERC-20 contract
     const currencyAddress = '0x3c499c542cEF5E3811e1192ce70d8cC03d5c3359'
     const salesContractAddress = '0xe65b75eb7c58ffc0bf0e671d64d0e1c6cd0d3e5b'
     const collectionAddress = '0xdeb398f41ccd290ee5114df7e498cf04fac916cb'
     const price = '20000'
+    const contractId = '674eb5613d739107bbd18ed2'
 
     const chainId = 137
 
@@ -358,8 +360,12 @@ export const Connected = () => {
       recipientAddress: address,
       currencyAddress,
       collectionAddress,
-      creditCardProviders: ['sardine'],
-      isDev: true,
+      creditCardProviders: ['sardine', 'transak'],
+      transakConfig: {
+        contractId,
+        apiKey: '5911d9ec-46b5-48fa-a755-d59a715ff0cf'
+      },
+      isDev: false,
       copyrightText: 'ⓒ2024 Sequence',
       onSuccess: (txnHash: string) => {
         console.log('success!', txnHash)