+++ packages/checkout/src/contexts/CheckoutModal.ts
@@ -16,6 +16,12 @@ interface OrderSummaryItem {
   tokenId: string
 }
 
+export interface TransakConfig {
+  apiKey: string
+  contractId: string
+  callDataOverride?: string
+}
+
 export interface CreditCardCheckout {
   chainId: number
   contractAddress: string
@@ -29,6 +35,8 @@ export interface CreditCardCheckout {
   nftQuantity: string
   nftDecimals?: string
   calldata: string
+  provider?: 'sardine' | 'transak'
+  transakConfig?: TransakConfig
   onSuccess?: (transactionHash: string, settings: CreditCardCheckout) => void
   onError?: (error: Error, settings: CreditCardCheckout) => void
   isDev?: boolean