+++ packages/kit/src/connectors/coinbaseWallet/CoinbaseWalletLogo.tsx
@@ -2,15 +2,15 @@ import React from 'react'
 
 import { LogoProps } from '../../types'
 
-export const CoinbaseWalletLogo: React.FunctionComponent = (props: LogoProps) => {
+export const CoinbaseWalletLogo: React.FunctionComponent<LogoProps> = props => {
   return (
-    <svg viewBox="0 0 28 28" fill="none" xmlns="http://www.w3.org/2000/svg" {...props}>
-      <rect width="28" height="28" fill="#2C5FF6" />
+    <svg xmlns="http://www.w3.org/2000/svg" viewBox="0 0 1024 1024" fill="none" {...props}>
+      <rect width="100%" height="100%" fill="#0052FF" rx="100" />
       <path
+        fill="#fff"
         fillRule="evenodd"
+        d="M152 512c0 198.823 161.177 360 360 360s360-161.177 360-360-161.177-360-360-360-360 161.177-360 360Zm268-116c-13.255 0-24 10.745-24 24v184c0 13.255 10.745 24 24 24h184c13.255 0 24-10.745 24-24V420c0-13.255-10.745-24-24-24H420Z"
         clipRule="evenodd"
-        d="M14 23.8C19.4124 23.8 23.8 19.4124 23.8 14C23.8 8.58761 19.4124 4.2 14 4.2C8.58761 4.2 4.2 8.58761 4.2 14C4.2 19.4124 8.58761 23.8 14 23.8ZM11.55 10.8C11.1358 10.8 10.8 11.1358 10.8 11.55V16.45C10.8 16.8642 11.1358 17.2 11.55 17.2H16.45C16.8642 17.2 17.2 16.8642 17.2 16.45V11.55C17.2 11.1358 16.8642 10.8 16.45 10.8H11.55Z"
-        fill="white"
       />
     </svg>
   )