+++ packages/kit/src/connectors/wagmiConnectors/sequenceConnector.ts
@@ -153,6 +153,8 @@ export function sequenceWallet(params: BaseSequenceConnectorOptions) {
           throw 'projectAccessKey not found'
         }
 
+        console.log('initWallet', projectAccessKey)
+
         const provider = sequence.initWallet(projectAccessKey, {
           defaultNetwork,
           transports: {