+++ packages/wallet/src/views/SendCollectible.tsx
@@ -200,7 +200,7 @@ export const SendCollectible = ({ chainId, contractAddress, tokenId }: SendColle
               toAddress,
               [tokenId],
               [ethers.toQuantity(sendAmount)],
-              []
+              new Uint8Array()
             ]) as `0x${string}`,
             gas: null
           },